perf(core): split GTRuntime from the GT admin API client
