+++ .size-limit.cjs
@@ -73,6 +73,7 @@ const reactNativeNode = (name, file) =>
 
 module.exports = [
   core('generaltranslation', 'index'),
+  core('generaltranslation/runtime', 'runtime'),
   core('generaltranslation/core', 'core'),
   core('generaltranslation/id', 'id'),
   core('generaltranslation/internal', 'internal'),
