+++ packages/react-core/src/hooks/useInternalLocaleSelector.ts
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react';
 import type { LocaleProperties } from '@generaltranslation/format/types';
 import { useCustomMapping, useLocales } from './i18n-config';
 import { useLocale } from './condition-store';
-import { getLocaleProperties } from 'generaltranslation';
+import { getLocaleProperties } from '@generaltranslation/format';
 
 // Explicit return type so the inferred type stays portable for downstream
 // packages (e.g. gt-react-native), otherwise the getLocaleProperties callback
